move fromHex method to test package
use bigFromHex for constants

fromHex retuns nil instead of panicking
