+++ bls12_381_test.go
@@ -2,6 +2,7 @@ package bls12381
 
 import (
 	"crypto/rand"
+	"encoding/hex"
 	"errors"
 	"flag"
 	"math/big"
@@ -38,4 +39,34 @@ func randScalars(max *big.Int, size int) []*big.Int {
 		scalars = append(scalars, a)
 	}
 	return scalars
+}
+
+func fromHex(size int, hexStrs ...string) []byte {
+	var out []byte
+	if size > 0 {
+		out = make([]byte, size*len(hexStrs))
+	}
+	for i := 0; i < len(hexStrs); i++ {
+		hexStr := hexStrs[i]
+		if hexStr[:2] == "0x" {
+			hexStr = hexStr[2:]
+		}
+		if len(hexStr)%2 == 1 {
+			hexStr = "0" + hexStr
+		}
+		bytes, err := hex.DecodeString(hexStr)
+		if err != nil {
+			return nil
+		}
+		if size <= 0 {
+			out = append(out, bytes...)
+		} else {
+			if len(bytes) > size {
+				return nil
+			}
+			offset := i*size + (size - len(bytes))
+			copy(out[offset:], bytes)
+		}
+	}
+	return out
 }
